Connected to Data of Materi
This commit has enabled the app to connect to all data related to Materi table. To see the way of using it, please refer to the MateriActivity in presentation/testactivity/MateriActivity.

There are also some additional changes in all the UIs that have been made. All UIs that have EditText should now call a method from class Helper, where it the method is trying to help to close keyboard view.