Fix on log in issue
It was known that when user logged in from other devices, it will cause confusion within the app where the token stored is doesn't match anymore with the token stored in the API database. In this update, HomeActivity will check logged in user's token through the API, and if it doesn't match then it will log out the user automatically.